+++ app/dashboard/vocabulary/[level]/word/page.tsx
@@ -6,22 +6,22 @@ import { useRouter } from "next/navigation";
 export default function WordPage() {
   const router = useRouter();
 
-  const dummyGrammars = [
-    { index: 0, word: "~恐れがある", furigana: "(~おそれがある)", mean: "“~할 우려가 있다.”, “~할 염려가 있다.”",
-      example: "地震による津波の恐れがありますので、高台に避難してください。", exMean: "지진으로 인해 쓰나미가 올 우려가 있으니, 높은 곳으로 대피하세요." },
-    { index: 1, word: "~べき", mean: "해야 한다.", example: "もっと努力すべきだ。", exMean: "더 노력해야 한다." },
+  const dummyWords = [
+    { index: 0, word: "即ち", furigana: "すなわち", mean: "즉" },
+    { index: 1, word: "勉強", furigana: "べんきょう", mean: "공부" },
+    { index: 2, word: "楽しい", furigana: "たのしい", mean: "즐겁다" },
   ];
 
   const [currentIndex, setCurrentIndex] = useState(0);
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   const [visibility, setVisibility] = useState({
-    explanation: false,
+    furigana: false,
     mean: false,
     workbook: false,
   });
 
-  const toggleVisibility = (key: "explanation" | "mean" | "workbook") => {
+  const toggleVisibility = (key: "furigana" | "mean" | "workbook") => {
     setVisibility((prev) => ({
       ...prev,
       [key]: !prev[key]
@@ -29,9 +29,9 @@ export default function WordPage() {
   };
 
   const handleNextWord = () => {
-    if (currentIndex < dummyGrammars.length - 1) {
+    if (currentIndex < dummyWords.length - 1) {
       setCurrentIndex((prevIndex) => prevIndex + 1);
-      setVisibility({ explanation: false, mean: false, workbook: false });
+      setVisibility({ furigana: false, mean: false, workbook: false });
     } else {
       setIsModalOpen(true); // 마지막 단어에서 모달 열기
     }
@@ -39,7 +39,7 @@ export default function WordPage() {
 
   const restartLearning = () => {
     setCurrentIndex(0);
-    setVisibility({ explanation: false, mean: false, workbook: false });
+    setVisibility({ furigana: false, mean: false, workbook: false });
     setIsModalOpen(false);
   };
 
@@ -67,14 +67,12 @@ export default function WordPage() {
             </button>
 
             {/* 본문 내용 */}
-            <p className="text-nihonred text-7xl font-bold">{dummyGrammars[currentIndex].word}</p>
-            <p className="text-nihonred text-7xl">{dummyGrammars[currentIndex].furigana}</p>
-            {visibility.mean && (
-              <p className="text-3xl font-semibold pt-4">{dummyGrammars[currentIndex].mean}</p>
+            {visibility.furigana && (
+              <p className="text-nihonred text-4xl">{dummyWords[currentIndex].furigana}</p>
             )}
-            <p className="text-2xl font-semibold pt-4 pl-4">{dummyGrammars[currentIndex].example}</p>
+            <p className="text-black text-8xl font-bold">{dummyWords[currentIndex].word}</p>
             {visibility.mean && (
-                <p className="text-xl">{dummyGrammars[currentIndex].exMean}</p>
+              <p className="text-nihonred text-4xl font-semibold pt-3">{dummyWords[currentIndex].mean}</p>
             )}
           </div>
 
@@ -95,8 +93,8 @@ export default function WordPage() {
               의미
             </button>
             <button className="w-[180px] h-[45px] font-bold border-2 border-nihonred rounded-lg"
-              onClick={() => toggleVisibility("explanation")}>
-              설명
+              onClick={() => toggleVisibility("furigana")}>
+              히라가나
             </button>
             <button className="w-[180px] h-[45px] bg-red-400 text-white font-bold rounded-lg"
               onClick={handleNextWord}>
