+++ app/layout.tsx
@@ -1,16 +1,31 @@
-import NavBar from "./components/NavBar";
+import type { Metadata } from "next";
+import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
   return (
-    <html lang="ko">
-      <body>
-        <div className="flex h-screen">
-          <NavBar />
-          <div className="flex-1 ml-[85px] xl:ml-64 min-h-screen overflow-y-auto">
-            {children}
-          </div>
-        </div>
+    <html lang="en">
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}> 
+        {children}
       </body>
     </html>
   );
