+++ app/page.tsx
@@ -2,5 +2,5 @@ import { redirect } from "next/navigation";
 
 export default function Home(): never {
   // 메인 화면 리다이렉션
-  redirect("/vocabulary");
+  redirect("/dashboard/vocabulary");
 }
